cleanup old upload code
